+++ src/middlewares/index.js
@@ -1,0 +1,13 @@
+const isAuthenticated = require("./isAuthenticated.middleware");
+const isVerified = require("./isVerified.middleware");
+const isActive = require("./isActive.middleware");
+const isNotDeleted = require("./isNotDeleted.middleware");
+
+// ** Need to be used in the same order!
+
+module.exports = {
+	isAuthenticated,
+	isVerified,
+	isActive,
+	isNotDeleted,
+};
